Correcao unidade

+++ database/migrations/2023_05_04_145855_create_itens_venda_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
 
             $table->string('tipo_unidade');
-            $table->integer('quantidade');
+            $table->double('quantidade', 6, 3);
             $table->decimal('preco');
 
             $table->foreignId('produto_id')->constrained()->onDelete('restrict');
